Make the GUI code modular.
